refactor(core,server): close the extension contract, move sandbox into it
PenguinContext and PenguinInterface were OPEN: the harness contributed its
members by augmenting @prismshadow/penguin-core/extension with `declare module`.
An extension type-checking against the contract therefore compiled against
members only one embedder supplies, and nothing at the core layer could tell
them apart from the contract's own. They are closed now and name every member.

The sandbox vocabulary moves into the contract, where a backend already looks:
SandboxPolicy, SandboxProvider, ConfinedArgv, the mode/dimension unions and the
registry/control surfaces are core's, and PenguinContext.sandbox /
PenguinInterface.sandbox are declared outright rather than merged in.

Only the vocabulary moved. SANDBOX_DIMENSIONS, providerDimensions and
requestedDimensions READ those shapes and are the embedder's, so they stay in
the server package. That is what keeps both extension subpaths types-only —
core's extension/index.js is still 33 bytes — so an extension package carries no
runtime dependency on either.

`terminals` cannot be in the contract: it is a server class. It is named on
HarnessContext, which extends the contract, so the emitted value still satisfies
it and the seam is unchanged. Reaching that member is now `ctx as
HarnessContext` — an explicit statement, at the point of use, that the extension
depends on this embedder. Closing it is what makes that cast necessary rather
than optional: the two call sites that read `terminals` off the bare contract
failed to compile until they said so.

Verified by injection rather than by absence of errors: reading `terminals` off
a plain PenguinContext from an extension package is TS2339; the same read behind
the cast compiles; the sandbox vocabulary still resolves through the server
subpath. Earlier probes also confirmed the cross-package checking is real —
misusing SandboxPolicy reports SandboxPolicy by name, so it resolves to a type
rather than degrading to any.
